yolov7/requirements.txt
# Usage: pip install -r requirements.txt





torch>=1.7.0
torchvision>=0.8.1
opencv-python>=4.1.2
PyYAML>=5.3.1
scipy==1.9.1
tqdm>=4.41.0
tensorboard>=2.4.1
seaborn==0.12.0
pandas>=1.1.4
matplotlib>=3.2.2



# Base ----------------------------------------
# matplotlib>=3.2.2
numpy==1.21.6
# opencv-python>=4.1.1
Pillow>=7.1.2
# PyYAML>=5.3.1
requests>=2.23.0
# scipy>=1.4.1
# torch>=1.7.0,!=1.12.0
# torchvision>=0.8.1,!=0.13.0
# tqdm>=4.41.0
protobuf<4.21.3

# Logging -------------------------------------
# tensorboard>=2.4.1
# wandb

# Plotting ------------------------------------
# pandas>=1.1.4
# seaborn>=0.11.0

# Export --------------------------------------
# coremltools>=4.1  # CoreML export
# onnx>=1.9.0  # ONNX export
# onnx-simplifier>=0.3.6  # ONNX simplifier
# scikit-learn==0.19.2  # CoreML quantization
# tensorflow>=2.4.1  # TFLite export
# tensorflowjs>=3.9.0  # TF.js export
# openvino-dev  # OpenVINO export

# Extras --------------------------------------
ipython  # interactive notebook
psutil  # system utilization
thop  # FLOPs computation
# albumentations>=1.0.3
# pycocotools>=2.0  # COCO mAP
# roboflow
